backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6
pillow>=9.0.0
python-dotenv>=1.0.0
numpy>=1.22.0
